Update HAL for latest version
